+++ src/session/clientsession.h
@@ -55,7 +55,7 @@ class ClientSession : public SocketSession {
 
   protected:
     void in_async_read();
-    void in_async_write(const std::string_view& data, int ack_count = 0);
+    void in_async_write(const std::string_view& data, size_t ack_count = 0);
     void out_async_read();
     void out_async_write(const std::string_view& data);
     void out_sent();
@@ -63,7 +63,7 @@ class ClientSession : public SocketSession {
     void udp_async_write(const std::string_view& data, const boost::asio::ip::udp::endpoint& endpoint);
     void udp_recv(const std::string_view& data, const boost::asio::ip::udp::endpoint& endpoint);
     void udp_sent();
-    void out_recv(const std::string_view& data, int ack_count = 0);
+    void out_recv(const std::string_view& data, size_t ack_count = 0);
 
     virtual void in_recv(const std::string_view& data);
     virtual void in_sent();
@@ -81,7 +81,7 @@ class ClientSession : public SocketSession {
     void start() override;
     void destroy(bool pipeline_call = false) override;
 
-    void recv_ack_cmd(int ack_count) override;
+    void recv_ack_cmd(size_t ack_count) override;
 };
 
 #endif // _CLIENTSESSION_H_
